+++ lxcraft/__init__.py
@@ -1,5 +1,6 @@
 from .debug import debug
 from .plan import Plan
-from .plan_element import PlanElement
+from .resource import Resource
+from .system import system
 
-__all__ = ["Plan", "PlanElement", "debug"]
+__all__ = ["Plan", "Resource", "debug", "system"]