+++ lxcraft/debian/apt_package.py
@@ -1,4 +1,3 @@
-import os
 from dataclasses import dataclass
 from subprocess import getstatusoutput
 
@@ -6,31 +5,35 @@ import lxcraft
 
 
 @dataclass
-class AptPackages(lxcraft.PlanElement):
+class AptPackages(lxcraft.Resource):
     """List of packages that must be installed"""
 
     package_list: list[str]
 
-    def get_actions(self):
-        self.missing_package_list = []
-        for package in self.package_list:
-            if self.is_installed(package):
-                continue
-            self.missing_package_list.append(package)
-        if self.missing_package_list:
-            return self.install_missing()
+    def create(self):
+        package_list = " ".join(self.package_list)
+        lxcraft.system(f"apt install -y {package_list}")
 
-    def install_missing(self):
-        package_list = " ".join(self.missing_package_list)
-        rc = os.system(f"apt install -y {package_list}")
-        if rc != 0:
-            raise Exception(f"Command terminated with non zero exit code {rc}")
+    def destroy(self):
+        package_list = " ".join(self.package_list)
+        lxcraft.system(f"apt remove -y {package_list}")
+
+    def is_created(self):
+        for package in self.package_list:
+            if not self.is_installed(package):
+                return False
+        return True
+
+    def is_consistent(self):
+        # TODO: Add support for version checking
+        return True
 
     @staticmethod
     def is_installed(package_name: str):
-        rc, _ = getstatusoutput(f"dpkg -s {package_name}")
+        command = (
+            f"dpkg-query --showformat {package_name}"
+            " '${Status}\n' --show passwd | grep -q '^install ok installed$'"
+        )
+        rc, _ = getstatusoutput(command)
+        lxcraft.debug("command", rc, command)
         return rc == 0
-
-    def destroy(self):
-        if self.missing_package_list:
-            os.system(f"apt remove -y {' '.join(self.missing_package_list)}")