+++ lxcraft/path/filecontent.py
@@ -8,7 +8,7 @@ import lxcraft
 
 
 @dataclass
-class FileContent(lxcraft.PlanElement):
+class FileContent(lxcraft.Resource):
     """Create file with content"""
 
     target_path: str
@@ -23,7 +23,28 @@ class FileContent(lxcraft.PlanElement):
     def set_templates_directory(directory: str):
         FileContent.templates_directory = directory
 
-    def get_source_text(self):
+    def create(self):
+        Path(self.target_path).write_text(self.get_template_text())
+        if self.owner_user != "" or self.owner_group != "":
+            self.chown()
+        if self.mode:
+            self.chmod()
+
+    def destroy(self):
+        Path(self.target_path).unlink()
+
+    def is_created(self):
+        return Path(self.target_path).exists()
+
+    def is_consistent(self):
+        return (
+            self.is_created()
+            and self.get_template_text() == Path(self.target_path).read_text()
+            and self.mode_matches()
+            and self.owner_matches()
+        )
+
+    def get_template_text(self):
         if self.source_path:
             source_path = Path(self.source_path)
         else:
@@ -39,45 +60,17 @@ class FileContent(lxcraft.PlanElement):
             replaced_text = replaced_text.replace(key, value)
         return replaced_text
 
-    def content_differs(self):
-        try:
-            target_text = Path(self.target_path).read_text()
-        except FileNotFoundError:
-            return True
-        return self.get_source_text() != target_text
-
-    def owner_differs(self):
-        if not self.owner_user and not self.owner_group:
-            return False
-
-        try:
-            current_uid = Path(self.target_path).stat().st_uid
-            current_gid = Path(self.target_path).stat().st_gid
-        except FileNotFoundError:  # The target file is not yet on the system
-            return True
+    def owner_matches(self):
+        current_uid = Path(self.target_path).stat().st_uid
+        current_gid = Path(self.target_path).stat().st_gid
         if self.owner_user and pwd.getpwnam(self.owner_user).pw_uid != current_uid:
-            return True
+            return False
         if self.owner_group and grp.getgrnam(self.owner_group).gr_gid != current_gid:
-            return True
+            return False
+        return True
 
-    def mode_differs(self):
-        try:
-            return self.mode != Path(self.target_path).stat().st_mode & 0o777
-        except FileNotFoundError:
-            return True
-
-    def get_actions(self):
-        return self.action_engine(
-            {
-                self.content_differs: self.create,
-                self.owner_differs: self.chown,
-                self.mode_differs: self.chmod,
-            }
-        )
-
-    def create(self):
-        lxcraft.debug("filecontent", "Creating", self.target_path)
-        Path(self.target_path).write_text(self.get_source_text())
+    def mode_matches(self):
+        return self.mode == Path(self.target_path).stat().st_mode & 0o777
 
     def chown(self):
         if self.owner_user == "":
@@ -92,6 +85,3 @@ class FileContent(lxcraft.PlanElement):
 
     def chmod(self):
         os.chmod(self.target_path, self.mode)
-
-    def destroy(self):
-        os.remove(self.target_path)