+++ lxcraft/python/pip_package.py
@@ -1,4 +1,3 @@
-import os
 from dataclasses import dataclass
 from subprocess import getstatusoutput
 
@@ -6,32 +5,30 @@ import lxcraft
 
 
 @dataclass
-class PipPackages(lxcraft.PlanElement):
+class PipPackages(lxcraft.Resource):
     """List of packages that must be installed"""
 
     package_list: list[str]
 
-    def get_actions(self):
-        self.missing_package_list = []
-        for package in self.package_list:
-            if self.is_installed(package):
-                continue
-            self.missing_package_list.append(package)
-        if self.missing_package_list:
-            return self.install_missing()
+    def create(self):
+        package_list = " ".join(self.package_list)
+        lxcraft.system(f"pip install {package_list}")
 
-    def install_missing(self):
-        package_list = " ".join(self.missing_package_list)
-        rc = os.system(f"pip install {package_list}")
-        if rc != 0:
-            raise Exception(f"Command terminated with non zero exit code {rc}")
+    def destroy(self):
+        package_list = " ".join(self.package_list)
+        lxcraft.system(f"pip uninstall -y {package_list}")
+
+    def is_created(self):
+        for package in self.package_list:
+            if not self.is_installed(package):
+                return False
+        return True
+
+    def is_consistent(self):
+        # TODO: Add support for version checking
+        return True
 
     @staticmethod
     def is_installed(package_name: str):
         rc, _ = getstatusoutput(f"pip show {package_name}")
         return rc == 0
-
-    def destroy(self):
-        package_list = " ".join(self.missing_package_list)
-        if package_list:
-            os.system(f"pip uninstall -y {package_list}")